modules: Better code for working out the dynamic module dir.
The current code used libdir from the pkgconfig file, but Debian
and Ubuntu do not put their modules inside this folder and thus it's
a bit flawed.

So take the static module dir and strip of the trailing parts that
might change with a new library version and make that tail end
dynamic. Should work on all platforms.